+++ src/main/java/com/ll/exam/article/service/ArticleService.java
@@ -32,7 +32,7 @@ public class ArticleService {
         return articleRepository.write(title,content,isBlind);
     }
 
-    public void modify(int id, String title, String content, boolean isBlind) {
+    public void modify(long id, String title, String content, boolean isBlind) {
         articleRepository.modify(id, title, content, isBlind);
     }
